fix: tipo de variable Long en lugar de Integer

+++ src/main/java/com/task/taskmanager/infrastructure/controller/StateController.java
@@ -68,7 +68,7 @@ public class StateController {
 	}
 
 	@DeleteMapping("/state/{id}")
-	public ResponseEntity<?> deleteState(@PathVariable("id") Integer id) {
+	public ResponseEntity<?> deleteState(@PathVariable("id") Long id) {
 		State state = stateService.getStateById(id);
 		if(state != null) {
 			stateService.deleteState(id);
