+++ src/main/java/com/task/taskmanager/infrastructure/service/spec/TaskService.java
@@ -8,8 +8,8 @@ public interface TaskService {
 	
 	public List<Task> getAllTask();
 	public Task createTask(Task task);
-	public Task getTaskById(Integer idTask);
+	public Task getTaskById(Long idTask);
 	public Task updateTask(Task task);
-	public boolean deleteTask(Integer idtasj);
+	public boolean deleteTask(Long idtasj);
 
 }
